scripts/clueless-mind/requirements.txt
pdfplumber>=0.11.0
PyPDF2>=3.0.0
PyGithub>=2.4.0
